+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New `forced` bool parameter to force a re-authentication when using `Authenticate()`.
+### Changed
+- `Authenticate()` now do not authenticate if `isAuthenticated` is already true and
+- `Authenticate()` also now blocks any authentication attempt if there is already an attempt going on.
+- `isAuthenticating` is also available for checking the authentication process.
 
 ## 0.3.1 - 2021-02-09
 ### Changed